refactor: Extract record stream management to separate class
- Remove the managing of the RecordStreamSource from BpmnAssert
