+++ src/main/java/io/camunda/testing/assertions/BpmnAssert.java
@@ -1,22 +1,13 @@
 package io.camunda.testing.assertions;
 
+import static io.camunda.testing.utils.RecordStreamManager.getRecordStreamSource;
+
 import io.camunda.zeebe.client.api.response.ActivatedJob;
 import io.camunda.zeebe.client.api.response.DeploymentEvent;
 import io.camunda.zeebe.client.api.response.ProcessInstanceEvent;
 import io.camunda.zeebe.client.api.response.PublishMessageResponse;
-import org.camunda.community.eze.RecordStreamSource;
 
 public abstract class BpmnAssert {
-
-  static ThreadLocal<RecordStreamSource> recordStreamSource = new ThreadLocal<>();
-
-  public static void init(final RecordStreamSource recordStreamSource) {
-    BpmnAssert.recordStreamSource.set(recordStreamSource);
-  }
-
-  public static void reset() {
-    recordStreamSource.remove();
-  }
 
   public static ProcessInstanceAssert assertThat(final ProcessInstanceEvent instanceEvent) {
     return new ProcessInstanceAssert(
@@ -33,14 +24,5 @@ public abstract class BpmnAssert {
 
   public static MessageAssert assertThat(final PublishMessageResponse publishMessageResponse) {
     return new MessageAssert(publishMessageResponse, getRecordStreamSource());
-  }
-
-  private static RecordStreamSource getRecordStreamSource() {
-    if (recordStreamSource.get() == null) {
-      throw new AssertionError(
-          "No RecordStreamSource is set. Please use @ZeebeAssertions extension and "
-              + "declare a RecordStreamSource field");
-    }
-    return recordStreamSource.get();
   }
 }
