+++ src/main/java/io/camunda/testing/utils/RecordStreamManager.java
@@ -1,0 +1,25 @@
+package io.camunda.testing.utils;
+
+import org.camunda.community.eze.RecordStreamSource;
+
+public abstract class RecordStreamManager {
+
+  static ThreadLocal<RecordStreamSource> recordStreamSource = new ThreadLocal<>();
+
+  public static void init(final RecordStreamSource recordStreamSource) {
+    RecordStreamManager.recordStreamSource.set(recordStreamSource);
+  }
+
+  public static void reset() {
+    recordStreamSource.remove();
+  }
+
+  public static RecordStreamSource getRecordStreamSource() {
+    if (recordStreamSource.get() == null) {
+      throw new AssertionError(
+          "No RecordStreamSource is set. Please use @ZeebeAssertions extension and "
+              + "declare a RecordStreamSource field");
+    }
+    return recordStreamSource.get();
+  }
+}
